RabbitMq/requirements.txt
python3-pika